@@ -15,44 +15,53 @@
     "url": "git://github.com/LibertyDSNP/parquetjs.git"
   },
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.575.0",
     "@types/long": "^4.0.2",
-    "@types/node-int64": "^0.4.29",
-    "@types/thrift": "^0.10.11",
+    "@types/node-int64": "^0.4.32",
+    "@types/thrift": "^0.10.17",
+    "@zenfs/core": "^0.15.1",
+    "brotli-wasm": "^3.0.0",
     "browserify-zlib": "^0.2.0",
-    "bson": "4.6.3",
-    "cross-fetch": "^3.1.4",
-    "int53": "^0.2.4",
-    "long": "^4.0.0",
-    "snappyjs": "^0.6.1",
-    "thrift": "0.16.0",
+    "bson": "6.7.0",
+    "int53": "^1.0.0",
+    "long": "^5.2.3",
+    "snappyjs": "^0.7.0",
+    "thrift": "0.20.0",
     "varint": "^6.0.0",
-    "wasm-brotli": "^2.0.2",
     "xxhash-wasm": "^1.0.2"
   },
   "devDependencies": {
-    "@types/chai": "^4.3.5",
-    "@types/json-schema": "^7.0.11",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.18.2",
-    "@types/sinon": "^10.0.15",
-    "@types/varint": "^6.0.1",
-    "assert": "^2.0.0",
-    "browserfs": "^1.4.3",
+    "@eslint/js": "^9.6.0",
+    "@smithy/util-stream": "^3.0.0",
+    "@types/chai": "^4.3.16",
+    "@types/eslint__js": "^8.42.3",
+    "@types/json-schema": "^7.0.15",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.12.12",
+    "@types/sinon": "^17.0.3",
+    "@types/varint": "^6.0.3",
+    "assert": "^2.1.0",
+    "aws-sdk-client-mock": "^4.0.0",
     "buffer": "^6.0.3",
-    "chai": "4.3.6",
-    "core-js": "^3.22.5",
-    "esbuild": "^0.19.2",
-    "mocha": "^10.2.0",
-    "msw": "^1.2.1",
+    "chai": "4.4.1",
+    "core-js": "^3.37.1",
+    "esbuild": "^0.21.2",
+    "eslint": "^8.57.0",
+    "eslint-plugin-mocha": "^10.4.3",
+    "events": "^3.3.0",
+    "mocha": "^10.4.0",
+    "msw": "^2.3.0",
     "object-stream": "^0.0.1",
+    "prettier": "3.3.2",
     "process": "^0.11.10",
-    "regenerator-runtime": "^0.13.11",
-    "sinon": "^15.1.0",
+    "regenerator-runtime": "^0.14.1",
+    "sinon": "^17.0.2",
     "sinon-chai": "^3.7.0",
     "sinon-chai-in-order": "^0.1.0",
     "stream-browserify": "^3.0.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
+    "tsx": "^4.10.2",
+    "typescript": "^5.5.2",
+    "typescript-eslint": "^7.14.1"
   },
   "scripts": {
     "build": "npm run build:node && npm run build:types",
@@ -60,9 +69,11 @@
     "build:node": "tsc -b",
     "build:browser": "node esbuild.js",
     "type": "tsc --noEmit",
-    "lint": "echo 'Linting, it is on the TODO list...'",
-    "test": "mocha -r ts-node/register 'test/{,!(browser)/**}/*.{js,ts}'",
-    "test:only": "mocha -r ts-node/register",
+    "lint": "eslint . && npx prettier . --check",
+    "lint:fix": "eslint --fix .",
+    "format": "npx prettier . --write",
+    "test": "mocha 'test/{,!(browser)/**}/*.{js,ts}'",
+    "test:only": "mocha",
     "clean": "rm -Rf ./dist",
     "prepublishOnly": "npm run clean && npm run build:node && npm run build:types && npm run build:browser",
     "thrift": "thrift -out gen-nodejs --gen js:ts parquet.thrift && thrift -out gen-nodejs --gen js:node parquet.thrift",
@@ -71,7 +82,7 @@
   "browser": {
     "assert": "assert",
     "events": "events",
-    "fs": "browserfs",
+    "fs": "@zenfs/core",
     "path": "path-browserify",
     "stream": "readable-stream",
     "thrift": "./node_modules/thrift/lib/nodejs/lib/thrift/browser.js",
