@@ -1,3 +1,3 @@
 // Generates the types for the es modules browser build
 
-export * from "../parquet";
+export * from '../parquet';
