@@ -1,3 +1,2 @@
-const buffer = require("buffer/").Buffer;
-export let Buffer =  buffer;
-
+const buffer = require('buffer/').Buffer;
+export let Buffer = buffer;
