@@ -1,7 +1,7 @@
-const express = require('express')
-const path = require("path")
-const app = express()
-const port = 3000
+const express = require('express');
+const path = require('path');
+const app = express();
+const port = 3000;
 
 app.use(express.static(path.join(__dirname, 'public')));
 app.engine('ejs', require('ejs').__express);
@@ -9,12 +9,12 @@ app.engine('ejs', require('ejs').__express);
 app.set('view engine', 'ejs');
 
 app.get('/', (req, res) => {
-    res.render('parquetFiles', {
-        title: "Parquet Files",
-        port: port
-    })
-})
+  res.render('parquetFiles', {
+    title: 'Parquet Files',
+    port: port,
+  });
+});
 
 app.listen(port, () => {
-    console.log(`Example app listening at http://localhost:${port}`)
-})
+  console.log(`Example app listening at http://localhost:${port}`);
+});
