@@ -1,30 +1,30 @@
-import { Statistics } from "../gen-nodejs/parquet_types"
-import { ParquetEnvelopeReader } from "./reader"
-import { FileMetaDataExt } from "./declare"
+import { Statistics } from '../gen-nodejs/parquet_types';
+import { ParquetEnvelopeReader } from './reader';
+import { FileMetaDataExt } from './declare';
 
 export interface BufferReaderOptions {
-  maxSpan?: number,
-  maxLength?: number,
-  queueWait?: number
+  maxSpan?: number;
+  maxLength?: number;
+  queueWait?: number;
   default_dictionary_size?: number;
-  metadata?: FileMetaDataExt
-  rawStatistics?: Statistics
+  metadata?: FileMetaDataExt;
+  rawStatistics?: Statistics;
 }
 
 interface BufferReaderQueueRow {
-  offset: number,
-  length: number,
-  resolve: (buf: Buffer) => void
-  reject: unknown
+  offset: number;
+  length: number;
+  resolve: (buf: Buffer) => void;
+  reject: unknown;
 }
 
 export default class BufferReader {
-  maxSpan: number
-  maxLength: number
-  queueWait: number
-  scheduled?: boolean
-  queue: Array<BufferReaderQueueRow>
-  envelopeReader: ParquetEnvelopeReader
+  maxSpan: number;
+  maxLength: number;
+  queueWait: number;
+  scheduled?: boolean;
+  queue: BufferReaderQueueRow[];
+  envelopeReader: ParquetEnvelopeReader;
 
   constructor(envelopeReader: ParquetEnvelopeReader, options: BufferReaderOptions) {
     options = options || {};
@@ -39,14 +39,14 @@ export default class BufferReader {
   read(offset: number, length: number): Promise<Buffer> {
     if (!this.scheduled) {
       this.scheduled = true;
-      setTimeout( () => {
+      setTimeout(() => {
         this.scheduled = false;
         this.processQueue();
-      },this.queueWait);
+      }, this.queueWait);
     }
 
-    return new Promise( (resolve, reject) => {
-      this.queue.push({offset,length,resolve,reject});
+    return new Promise((resolve, reject) => {
+      this.queue.push({ offset, length, resolve, reject });
     });
   }
 
@@ -54,9 +54,9 @@ export default class BufferReader {
     const queue = this.queue;
     if (!queue.length) return;
     this.queue = [];
-    queue.sort( (a,b) => a.offset - b.offset);
+    queue.sort((a, b) => a.offset - b.offset);
 
-    var subqueue: Array<BufferReaderQueueRow> = [];
+    let subqueue: BufferReaderQueueRow[] = [];
 
     const readSubqueue = async () => {
       if (!subqueue.length) {
@@ -66,21 +66,21 @@ export default class BufferReader {
       const processQueue = subqueue;
       subqueue = [];
 
-      const lastElement = processQueue[processQueue.length-1];
+      const lastElement = processQueue[processQueue.length - 1];
       const start = processQueue[0].offset;
-      const finish = lastElement.offset +lastElement.length;
+      const finish = lastElement.offset + lastElement.length;
       const buffer = await this.envelopeReader.readFn(start, finish - start);
 
-      processQueue.forEach(async d => {
-        d.resolve(buffer.slice(d.offset - start, d.offset + d.length - start));
+      processQueue.forEach(async (d) => {
+        d.resolve(buffer.subarray(d.offset - start, d.offset + d.length - start));
       });
     };
 
-    queue.forEach((d,i) => {
-      const prev = queue[i-1];
-      if (!prev || (d.offset - (prev.offset + prev.length)) < this.maxSpan) {
+    queue.forEach((d, i) => {
+      const prev = queue[i - 1];
+      if (!prev || d.offset - (prev.offset + prev.length) < this.maxSpan) {
         subqueue.push(d);
-        if ( (d.offset + d.length) - subqueue[0].offset > this.maxLength) {
+        if (d.offset + d.length - subqueue[0].offset > this.maxLength) {
           readSubqueue();
         }
       } else {
@@ -90,4 +90,4 @@ export default class BufferReader {
     });
     readSubqueue();
   }
-};
+}
