@@ -1,5 +1,4 @@
-export * as PLAIN from './plain'
-export * as RLE from './rle'
-export * as PLAIN_DICTIONARY from './plain_dictionary'
-
-
+export * as PLAIN from './plain';
+export * as RLE from './rle';
+export * as PLAIN_DICTIONARY from './plain_dictionary';
+export * as RLE_DICTIONARY from './plain_dictionary';
