@@ -1,8 +1,8 @@
-import INT53 from "int53";
-import { Cursor, Options } from "./types";
+import INT53 from 'int53';
+import { Cursor, Options } from './types';
 
-function encodeValues_BOOLEAN(values: Array<boolean>) {
-  let buf = Buffer.alloc(Math.ceil(values.length / 8));
+function encodeValues_BOOLEAN(values: boolean[]) {
+  const buf = Buffer.alloc(Math.ceil(values.length / 8));
   buf.fill(0);
 
   for (let i = 0; i < values.length; ++i) {
@@ -15,10 +15,10 @@ function encodeValues_BOOLEAN(values: Array<boolean>) {
 }
 
 function decodeValues_BOOLEAN(cursor: Cursor, count: number) {
-  let values = [];
+  const values = [];
 
   for (let i = 0; i < count; ++i) {
-    let b = cursor.buffer[cursor.offset + Math.floor(i / 8)];
+    const b = cursor.buffer[cursor.offset + Math.floor(i / 8)];
     values.push((b & (1 << i % 8)) > 0);
   }
 
@@ -26,13 +26,13 @@ function decodeValues_BOOLEAN(cursor: Cursor, count: number) {
   return values;
 }
 
-function encodeValues_INT32(values: Array<number>, opts: Options) {
+function encodeValues_INT32(values: number[], opts: Options) {
   const isDecimal = opts?.originalType === 'DECIMAL' || opts?.column?.originalType === 'DECIMAL';
   const scale = opts?.scale || 0;
-  let buf = Buffer.alloc(4 * values.length);
+  const buf = Buffer.alloc(4 * values.length);
   for (let i = 0; i < values.length; i++) {
     if (isDecimal) {
-      buf.writeInt32LE(values[i] *  Math.pow(10, scale), i * 4);
+      buf.writeInt32LE(values[i] * Math.pow(10, scale), i * 4);
     } else {
       buf.writeInt32LE(values[i], i * 4);
     }
@@ -61,13 +61,13 @@ function decodeValues_INT32(cursor: Cursor, count: number, opts: Options) {
   return values;
 }
 
-function encodeValues_INT64(values: Array<number>, opts: Options) {
+function encodeValues_INT64(values: number[], opts: Options) {
   const isDecimal = opts?.originalType === 'DECIMAL' || opts?.column?.originalType === 'DECIMAL';
   const scale = opts?.scale || 0;
-  let buf = Buffer.alloc(8 * values.length);
+  const buf = Buffer.alloc(8 * values.length);
   for (let i = 0; i < values.length; i++) {
     if (isDecimal) {
-      buf.writeBigInt64LE(BigInt(Math.floor(values[i] *  Math.pow(10, scale))), i * 8);
+      buf.writeBigInt64LE(BigInt(Math.floor(values[i] * Math.pow(10, scale))), i * 8);
     } else {
       buf.writeBigInt64LE(BigInt(values[i]), i * 8);
     }
@@ -81,7 +81,7 @@ function decodeValues_INT64(cursor: Cursor, count: number, opts: Options) {
   const name = opts.name || opts.column?.name || undefined;
   try {
     if (opts.originalType === 'DECIMAL' || opts.column?.originalType === 'DECIMAL') {
-      let columnOptions: any = opts.column?.originalType ? opts.column : opts;
+      const columnOptions: any = opts.column?.originalType ? opts.column : opts;
       values = decodeValues_DECIMAL(cursor, count, columnOptions);
     } else {
       for (let i = 0; i < count; ++i) {
@@ -102,12 +102,12 @@ function decodeValues_DECIMAL(cursor: Cursor, count: number, opts: Options) {
   // Default scale to 0 per spec
   const scale = opts.scale || 0;
 
-  const name = opts.name || undefined
+  const name = opts.name || undefined;
   if (!precision) {
     throw `missing option: precision (required for DECIMAL) for column: ${name}`;
   }
 
-  let values = [];
+  const values = [];
 
   // by default we prepare the offset and bufferFunction to work with 32bit integers
   let offset = 4;
@@ -118,7 +118,7 @@ function decodeValues_DECIMAL(cursor: Cursor, count: number, opts: Options) {
     bufferFunction = (offset: number) => cursor.buffer.readBigInt64LE(offset);
   }
   for (let i = 0; i < count; ++i) {
-    const bufferSize = cursor.size || 0
+    const bufferSize = cursor.size || 0;
     if (bufferSize === 0 || cursor.offset < bufferSize) {
       const fullValue = bufferFunction(cursor.offset);
       const valueWithDecimalApplied = Number(fullValue) / Math.pow(10, scale);
@@ -130,8 +130,8 @@ function decodeValues_DECIMAL(cursor: Cursor, count: number, opts: Options) {
   return values;
 }
 
-function encodeValues_INT96(values: Array<number>) {
-  let buf = Buffer.alloc(12 * values.length);
+function encodeValues_INT96(values: number[]) {
+  const buf = Buffer.alloc(12 * values.length);
 
   for (let i = 0; i < values.length; i++) {
     if (values[i] >= 0) {
@@ -146,9 +146,8 @@ function encodeValues_INT96(values: Array<number>) {
   return buf;
 }
 
-function decodeValues_INT96(cursor: Cursor, count: number, opts: Options) {
-  let values = [];
-  
+function decodeValues_INT96(cursor: Cursor, count: number) {
+  const values = [];
   for (let i = 0; i < count; ++i) {
     // Read the low and high parts as BigInts
     const lowBigInt = BigInt(INT53.readInt64LE(cursor.buffer, cursor.offset));
@@ -158,15 +157,15 @@ function decodeValues_INT96(cursor: Cursor, count: number, opts: Options) {
     const value = lowBigInt | highBigInt;
 
     values.push(value);
-    
+
     cursor.offset += 12;
   }
 
   return values;
 }
 
-function encodeValues_FLOAT(values: Array<number>) {
-  let buf = Buffer.alloc(4 * values.length);
+function encodeValues_FLOAT(values: number[]) {
+  const buf = Buffer.alloc(4 * values.length);
   for (let i = 0; i < values.length; i++) {
     buf.writeFloatLE(values[i], i * 4);
   }
@@ -175,7 +174,7 @@ function encodeValues_FLOAT(values: Array<number>) {
 }
 
 function decodeValues_FLOAT(cursor: Cursor, count: number) {
-  let values = [];
+  const values = [];
 
   for (let i = 0; i < count; ++i) {
     values.push(cursor.buffer.readFloatLE(cursor.offset));
@@ -185,8 +184,8 @@ function decodeValues_FLOAT(cursor: Cursor, count: number) {
   return values;
 }
 
-function encodeValues_DOUBLE(values: Array<number>) {
-  let buf = Buffer.alloc(8 * values.length);
+function encodeValues_DOUBLE(values: number[]) {
+  const buf = Buffer.alloc(8 * values.length);
   for (let i = 0; i < values.length; i++) {
     buf.writeDoubleLE(values[i], i * 8);
   }
@@ -195,7 +194,7 @@ function encodeValues_DOUBLE(values: Array<number>) {
 }
 
 function decodeValues_DOUBLE(cursor: Cursor, count: number) {
-  let values = [];
+  const values = [];
 
   for (let i = 0; i < count; ++i) {
     values.push(cursor.buffer.readDoubleLE(cursor.offset));
@@ -205,16 +204,15 @@ function decodeValues_DOUBLE(cursor: Cursor, count: number) {
   return values;
 }
 
-// Waylands reminder to check again
-function encodeValues_BYTE_ARRAY(values: Array<Uint8Array>) {
+function encodeValues_BYTE_ARRAY(values: Uint8Array[]) {
   let buf_len = 0;
-  const returnedValues: Array<Buffer> = [];
+  const returnedValues: Buffer[] = [];
   for (let i = 0; i < values.length; i++) {
     returnedValues[i] = Buffer.from(values[i]);
     buf_len += 4 + returnedValues[i].length;
   }
 
-  let buf = Buffer.alloc(buf_len);
+  const buf = Buffer.alloc(buf_len);
   let buf_pos = 0;
   for (let i = 0; i < returnedValues.length; i++) {
     buf.writeUInt32LE(returnedValues[i].length, buf_pos);
@@ -226,12 +224,12 @@ function encodeValues_BYTE_ARRAY(values: Array<Uint8Array>) {
 }
 
 function decodeValues_BYTE_ARRAY(cursor: Cursor, count: number) {
-  let values = [];
+  const values = [];
 
   for (let i = 0; i < count; ++i) {
-    let len = cursor.buffer.readUInt32LE(cursor.offset);
+    const len = cursor.buffer.readUInt32LE(cursor.offset);
     cursor.offset += 4;
-    values.push(cursor.buffer.slice(cursor.offset, cursor.offset + len));
+    values.push(cursor.buffer.subarray(cursor.offset, cursor.offset + len));
     cursor.offset += len;
   }
 
@@ -256,34 +254,28 @@ function byteArrayToDecimal(bytes: Uint8Array, scale: number): string {
   return `${integerPart}.${fractionalPart.toString().padStart(scale, '0')}`;
 }
 
-function encodeValues_FIXED_LEN_BYTE_ARRAY(
-  values: Array<Uint8Array>,
-  opts: Options
-) {
+function encodeValues_FIXED_LEN_BYTE_ARRAY(values: Uint8Array[], opts: Options) {
   if (!opts.typeLength) {
-    throw "missing option: typeLength (required for FIXED_LEN_BYTE_ARRAY)";
+    throw 'missing option: typeLength (required for FIXED_LEN_BYTE_ARRAY)';
   }
 
-  const returnedValues: Array<Buffer> = [];
+  const returnedValues: Buffer[] = [];
   for (let i = 0; i < values.length; i++) {
     returnedValues[i] = Buffer.from(values[i]);
 
     if (returnedValues[i].length !== opts.typeLength) {
-      throw "invalid value for FIXED_LEN_BYTE_ARRAY: " + returnedValues[i];
+      throw 'invalid value for FIXED_LEN_BYTE_ARRAY: ' + returnedValues[i];
     }
   }
 
   return Buffer.concat(returnedValues);
 }
 
-function decodeValues_FIXED_LEN_BYTE_ARRAY(
-  cursor: Cursor,
-  count: number,
-  opts: Options
-) {
-
-  if (!opts.typeLength) {
-    throw "missing option: typeLength (required for FIXED_LEN_BYTE_ARRAY)";
+function decodeValues_FIXED_LEN_BYTE_ARRAY(cursor: Cursor, count: number, opts: Options) {
+  const values = [];
+  const typeLength = opts.typeLength ?? (opts.column ? opts.column.typeLength : undefined);
+  if (!typeLength) {
+    throw 'missing option: typeLength (required for FIXED_LEN_BYTE_ARRAY)';
   }
 
   const isDecimal = opts?.originalType === 'DECIMAL' || opts?.column?.originalType === 'DECIMAL';
@@ -297,89 +289,82 @@ function decodeValues_FIXED_LEN_BYTE_ARRAY(
     return returnedValues;
   }
 
-  const values = [];
   for (let i = 0; i < count; ++i) {
-    values.push(
-      cursor.buffer.slice(cursor.offset, cursor.offset + opts.typeLength)
-    );
-    cursor.offset += opts.typeLength;
+    values.push(cursor.buffer.subarray(cursor.offset, cursor.offset + typeLength));
+    cursor.offset += typeLength;
   }
 
   return values;
 }
 
-type ValidValueTypes = "BOOLEAN" | "INT32" | "INT64" | "INT96" | "FLOAT" | "DOUBLE" | "BYTE_ARRAY" | "FIXED_LEN_BYTE_ARRAY"
+type ValidValueTypes =
+  | 'BOOLEAN'
+  | 'INT32'
+  | 'INT64'
+  | 'INT96'
+  | 'FLOAT'
+  | 'DOUBLE'
+  | 'BYTE_ARRAY'
+  | 'FIXED_LEN_BYTE_ARRAY';
 
-export const encodeValues = function (
-  type: ValidValueTypes | string,
-  values: Array<unknown>,
-  opts: Options
-) {
+export const encodeValues = function (type: ValidValueTypes | string, values: unknown[], opts: Options) {
   switch (type) {
-    case "BOOLEAN":
-      return encodeValues_BOOLEAN(values as Array<boolean>);
+    case 'BOOLEAN':
+      return encodeValues_BOOLEAN(values as boolean[]);
 
-    case "INT32":
-      return encodeValues_INT32(values as Array<number>, opts);
+    case 'INT32':
+      return encodeValues_INT32(values as number[], opts);
 
-    case "INT64":
-      return encodeValues_INT64(values as Array<number>, opts);
+    case 'INT64':
+      return encodeValues_INT64(values as number[], opts);
 
-    case "INT96":
-      return encodeValues_INT96(values as Array<number>);
+    case 'INT96':
+      return encodeValues_INT96(values as number[]);
 
-    case "FLOAT":
-      return encodeValues_FLOAT(values as Array<number>);
+    case 'FLOAT':
+      return encodeValues_FLOAT(values as number[]);
 
-    case "DOUBLE":
-      return encodeValues_DOUBLE(values as Array<number>);
+    case 'DOUBLE':
+      return encodeValues_DOUBLE(values as number[]);
 
-    case "BYTE_ARRAY":
-      return encodeValues_BYTE_ARRAY(values as Array<Uint8Array>);
+    case 'BYTE_ARRAY':
+      return encodeValues_BYTE_ARRAY(values as Uint8Array[]);
 
-    case "FIXED_LEN_BYTE_ARRAY":
-      return encodeValues_FIXED_LEN_BYTE_ARRAY(
-        values as Array<Uint8Array>,
-        opts
-      );
+    case 'FIXED_LEN_BYTE_ARRAY':
+      return encodeValues_FIXED_LEN_BYTE_ARRAY(values as Uint8Array[], opts);
 
     default:
-      throw "unsupported type: " + type;
+      throw 'unsupported type: ' + type;
   }
 };
 
-export const decodeValues = function (
-  type: ValidValueTypes | string,
-  cursor: Cursor,
-  count: number,
-  opts: Options
-) {
+export const decodeValues = function (type: ValidValueTypes | string, cursor: Cursor, count: number, opts: Options) {
   switch (type) {
-    case "BOOLEAN":
+    case 'BOOLEAN':
       return decodeValues_BOOLEAN(cursor, count);
 
-    case "INT32":
+    case 'INT32':
       return decodeValues_INT32(cursor, count, opts);
 
-    case "INT64":
+    case 'INT64':
       return decodeValues_INT64(cursor, count, opts);
 
-    case "INT96":
-      return decodeValues_INT96(cursor, count, opts);
+    case 'INT96':
+      return decodeValues_INT96(cursor, count);
 
-    case "FLOAT":
+    case 'FLOAT':
       return decodeValues_FLOAT(cursor, count);
 
-    case "DOUBLE":
+    case 'DOUBLE':
       return decodeValues_DOUBLE(cursor, count);
 
-    case "BYTE_ARRAY":
+    case 'BYTE_ARRAY':
       return decodeValues_BYTE_ARRAY(cursor, count);
 
-    case "FIXED_LEN_BYTE_ARRAY":
+    case 'FIXED_LEN_BYTE_ARRAY':
       return decodeValues_FIXED_LEN_BYTE_ARRAY(cursor, count, opts);
 
     default:
-      throw "unsupported type: " + type;
+      throw 'unsupported type: ' + type;
   }
 };
