@@ -6,9 +6,8 @@ import * as parquet_schema from './schema';
 import * as parquet_codec from './codec';
 import * as parquet_compression from './compression';
 import * as parquet_types from './types';
-import BufferReader, {BufferReaderOptions} from './bufferReader';
+import BufferReader, { BufferReaderOptions } from './bufferReader';
 import * as bloomFilterReader from './bloomFilterIO/bloomFilterReader';
-import fetch from 'cross-fetch';
 import {
   ParquetCodec,
   Parameter,
@@ -22,13 +21,14 @@ import {
   FileMetaDataExt,
   NewPageHeader,
   RowGroupExt,
-  ColumnChunkExt
+  ColumnChunkExt,
 } from './declare';
-import {Cursor, Options} from './codec/types';
+import { Cursor, Options } from './codec/types';
+import { GetObjectCommand, HeadObjectCommand, S3Client } from '@aws-sdk/client-s3';
+import type { Readable } from 'stream';
+import type { Blob } from 'buffer';
 
-const {
-  getBloomFiltersFor,
-} = bloomFilterReader;
+const { getBloomFiltersFor } = bloomFilterReader;
 
 /**
  * Parquet File Magic String
@@ -50,12 +50,11 @@ const PARQUET_RDLVL_ENCODING = 'RLE';
  * A parquet cursor is used to retrieve rows from a parquet file in order
  */
 class ParquetCursor {
-
   metadata: FileMetaDataExt;
   envelopeReader: ParquetEnvelopeReader;
   schema: parquet_schema.ParquetSchema;
-  columnList: Array<Array<unknown>>;
-  rowGroup: Array<unknown>;
+  columnList: unknown[][];
+  rowGroup: unknown[];
   rowGroupIndex: number;
   cursorIndex: number;
 
@@ -65,7 +64,12 @@ class ParquetCursor {
    * advanced and internal use cases. Consider using getCursor() on the
    * ParquetReader instead
    */
-  constructor(metadata: FileMetaDataExt, envelopeReader: ParquetEnvelopeReader, schema: parquet_schema.ParquetSchema, columnList: Array<Array<unknown>>) {
+  constructor(
+    metadata: FileMetaDataExt,
+    envelopeReader: ParquetEnvelopeReader,
+    schema: parquet_schema.ParquetSchema,
+    columnList: unknown[][]
+  ) {
     this.metadata = metadata;
     this.envelopeReader = envelopeReader;
     this.schema = schema;
@@ -82,14 +86,14 @@ class ParquetCursor {
   async next() {
     if (this.cursorIndex >= this.rowGroup.length) {
       if (this.rowGroupIndex >= this.metadata.row_groups.length) {
-
         return null;
       }
 
-      let rowBuffer = await this.envelopeReader.readRowGroup(
+      const rowBuffer = await this.envelopeReader.readRowGroup(
         this.schema,
         this.metadata.row_groups[this.rowGroupIndex],
-        this.columnList);
+        this.columnList
+      );
 
       this.rowGroup = parquet_shredder.materializeRecords(this.schema, rowBuffer);
       this.rowGroupIndex++;
@@ -107,7 +111,7 @@ class ParquetCursor {
     this.rowGroupIndex = 0;
     this.cursorIndex = 0;
   }
-};
+}
 
 /**
  * A parquet reader allows retrieving the rows from a parquet file in order.
@@ -117,33 +121,42 @@ class ParquetCursor {
  * avoid leaking file descriptors.
  */
 export class ParquetReader {
-
   envelopeReader: ParquetEnvelopeReader | null;
   metadata: FileMetaDataExt | null;
-  schema: parquet_schema.ParquetSchema
+  schema: parquet_schema.ParquetSchema;
 
   /**
    * Open the parquet file pointed to by the specified path and return a new
    * parquet reader
    */
   static async openFile(filePath: string | Buffer | URL, options?: BufferReaderOptions) {
-    let envelopeReader = await ParquetEnvelopeReader.openFile(filePath, options);
+    const envelopeReader = await ParquetEnvelopeReader.openFile(filePath, options);
     return this.openEnvelopeReader(envelopeReader, options);
   }
 
   static async openBuffer(buffer: Buffer, options?: BufferReaderOptions) {
-    let envelopeReader = await ParquetEnvelopeReader.openBuffer(buffer, options);
+    const envelopeReader = await ParquetEnvelopeReader.openBuffer(buffer, options);
     return this.openEnvelopeReader(envelopeReader, options);
   }
 
   /**
-   * Open the parquet file from S3 using the supplied aws client and params
-   * The params have to include `Bucket` and `Key` to the file requested
-   * This function returns a new parquet reader
+   * Open the parquet file from S3 using the supplied aws client [, commands] and params
+   * The params have to include `Bucket` and `Key` to the file requested,
+   * If using v3 of the AWS SDK, combine the client and commands into an object wiht keys matching
+   * the original module names, and do not instantiate the commands; pass them as classes/modules.
+   *
+   * This function returns a new parquet reader [ or throws an Error.]
    */
-  static async openS3(client: ClientS3, params: ClientParameters, options?: BufferReaderOptions) {
-    let envelopeReader = await ParquetEnvelopeReader.openS3(client, params, options);
-    return this.openEnvelopeReader(envelopeReader, options);
+  static async openS3(client: any, params: ClientParameters, options?: BufferReaderOptions) {
+    try {
+      const envelopeReader: ParquetEnvelopeReader =
+        'function' === typeof client['headObject']
+          ? await ParquetEnvelopeReader.openS3(client as ClientS3, params, options) // S3 client v2
+          : await ParquetEnvelopeReader.openS3v3(client as S3Client, params, options); // S3 client v3
+      return this.openEnvelopeReader(envelopeReader, options);
+    } catch (e: any) {
+      throw new Error(`Error accessing S3 Bucket ${params.Bucket}. Message: ${e.message}`);
+    }
   }
 
   /**
@@ -153,7 +166,7 @@ export class ParquetReader {
    * This function returns a new parquet reader
    */
   static async openUrl(params: Parameter | URL | string, options?: BufferReaderOptions) {
-    let envelopeReader = await ParquetEnvelopeReader.openUrl(params, options);
+    const envelopeReader = await ParquetEnvelopeReader.openUrl(params, options);
     return this.openEnvelopeReader(envelopeReader, options);
   }
 
@@ -164,7 +177,7 @@ export class ParquetReader {
     try {
       await envelopeReader.readHeader();
 
-      let metadata = await envelopeReader.readFooter();
+      const metadata = await envelopeReader.readFooter();
 
       return new ParquetReader(metadata, envelopeReader, opts);
     } catch (err) {
@@ -181,15 +194,16 @@ export class ParquetReader {
    */
   constructor(metadata: FileMetaDataExt, envelopeReader: ParquetEnvelopeReader, opts?: BufferReaderOptions) {
     opts = opts || {};
+
     if (!PARQUET_VERSIONS.includes(metadata.version)) {
       throw 'invalid parquet version';
     }
 
     // If metadata is a json file then we need to convert INT64 and CTIME
     if (metadata.json) {
-      const convert = (o: {[string: string]: any } ) => {
-        if (o &&  typeof o === 'object') {
-          Object.keys(o).forEach(key => o[key] = convert(o[key]));
+      const convert = (o: Record<string, any>) => {
+        if (o && typeof o === 'object') {
+          Object.keys(o).forEach((key) => (o[key] = convert(o[key])));
           if (o.parquetType === 'CTIME') {
             return new Date(o.value);
           } else if (o.parquetType === 'INT64') {
@@ -200,14 +214,16 @@ export class ParquetReader {
       };
 
       // Go through all PageLocation objects and set the proper prototype
-      metadata.row_groups.forEach(rowGroup => {
-        rowGroup.columns.forEach(column => {
+      metadata.row_groups.forEach((rowGroup) => {
+        rowGroup.columns.forEach((column) => {
           if (column.offsetIndex) {
-            Promise.resolve(column.offsetIndex).then(offset => (offset.page_locations.forEach(d => {
-              if (Array.isArray(d)) {
-                Object.setPrototypeOf(d,parquet_thrift.PageLocation.prototype);
-              }
-            })));
+            Promise.resolve(column.offsetIndex).then((offset) =>
+              offset.page_locations.forEach((d) => {
+                if (Array.isArray(d)) {
+                  Object.setPrototypeOf(d, parquet_thrift.PageLocation.prototype);
+                }
+              })
+            );
           }
         });
       });
@@ -218,21 +234,23 @@ export class ParquetReader {
     this.metadata = envelopeReader.metadata = metadata;
     this.envelopeReader = envelopeReader;
     this.schema = envelopeReader.schema = new parquet_schema.ParquetSchema(
-        decodeSchema(
-            this.metadata.schema.slice(1)) as SchemaDefinition);
+      decodeSchema(this.metadata.schema.slice(1)) as SchemaDefinition
+    );
 
     /* decode any statistics values */
     if (this.metadata.row_groups && !this.metadata.json && !opts.rawStatistics) {
-      this.metadata.row_groups.forEach(row => row.columns.forEach( col => {
-        const stats = col.meta_data!.statistics;
-        if (stats) {
-          const field = this.schema.findField(col.meta_data!.path_in_schema);
-          stats.max_value = decodeStatisticsValue(stats.max_value, field);
-          stats.min_value = decodeStatisticsValue(stats.min_value, field);
-          stats.min = decodeStatisticsValue(stats.min, field);
-          stats.max = decodeStatisticsValue(stats.max, field);
-        }
-      }));
+      this.metadata.row_groups.forEach((row) =>
+        row.columns.forEach((col) => {
+          const stats = col.meta_data!.statistics;
+          if (stats) {
+            const field = this.schema.findField(col.meta_data!.path_in_schema);
+            stats.max_value = decodeStatisticsValue(stats.max_value, field);
+            stats.min_value = decodeStatisticsValue(stats.min_value, field);
+            stats.min = decodeStatisticsValue(stats.min, field);
+            stats.max = decodeStatisticsValue(stats.max, field);
+          }
+        })
+      );
     }
   }
 
@@ -246,15 +264,13 @@ export class ParquetReader {
    *  }
    * ```
    */
-  async* [Symbol.asyncIterator]() {
+  async *[Symbol.asyncIterator]() {
     const cursor = this.getCursor();
     let record = null;
-    while (record = await cursor.next()) {
+    while ((record = await cursor.next())) {
       yield record;
     }
-}
-
-
+  }
 
   /**
    * Return a cursor to the file. You may open more than one cursor and use
@@ -265,25 +281,21 @@ export class ParquetReader {
    * from disk. An empty array or no value implies all columns. A list of column
    * names means that only those columns should be loaded from disk.
    */
-  getCursor(columnList?: Array<Array<unknown>>) {
+  getCursor(columnList?: unknown[][]) {
     if (!columnList) {
       columnList = [];
     }
 
-    columnList = columnList.map((x: Array<unknown>) => x.constructor === Array ? x : [x]);
+    columnList = columnList.map((x: unknown[]) => (x.constructor === Array ? x : [x]));
 
-    return new ParquetCursor(
-        this.metadata!,
-        this.envelopeReader!,
-        this.schema,
-        columnList);
+    return new ParquetCursor(this.metadata!, this.envelopeReader!, this.schema, columnList);
   }
 
   async getBloomFiltersFor(columnNames: string[]) {
     const bloomFilterData = await getBloomFiltersFor(columnNames, this.envelopeReader!);
     return bloomFilterData.reduce((acc: Record<string, typeof bloomFilterData>, value) => {
-      if (acc[value.columnName]) acc[value.columnName].push(value)
-      else acc[value.columnName] = [value]
+      if (acc[value.columnName]) acc[value.columnName].push(value);
+      else acc[value.columnName] = [value];
       return acc;
     }, {});
   }
@@ -307,8 +319,8 @@ export class ParquetReader {
    * Returns the user (key/value) metadata for this file
    */
   getMetadata() {
-    let md: Record<string, unknown> = {};
-    for (let kv of this.metadata!.key_value_metadata!) {
+    const md: Record<string, unknown> = {};
+    for (const kv of this.metadata!.key_value_metadata!) {
       md[kv.key] = kv.value;
     }
 
@@ -316,18 +328,19 @@ export class ParquetReader {
   }
 
   async exportMetadata(indent: string | number | undefined) {
-    function replacer(_key: unknown, value: parquet_thrift.PageLocation | bigint | {[string:string]: any}) {
+    function replacer(_key: unknown, value: parquet_thrift.PageLocation | bigint | Record<string, any>) {
       if (value instanceof parquet_thrift.PageLocation) {
         return [value.offset, value.compressed_page_size, value.first_row_index];
       }
 
       if (typeof value === 'object') {
-        for (let k in value) {
+        for (const k in value) {
           if (value[k] instanceof Date) {
-            value[k].toJSON = () => ({
-              parquetType: 'CTIME',
-              value: value[k].valueOf()
-            });
+            value[k].toJSON = () =>
+              JSON.stringify({
+                parquetType: 'CTIME',
+                value: value[k].valueOf(),
+              });
           }
         }
       }
@@ -342,29 +355,29 @@ export class ParquetReader {
         } else {
           return {
             parquetType: 'INT64',
-            value: [...value.buffer]
+            value: [...value.buffer],
           };
         }
       } else {
         return value;
       }
     }
-    const metadata = Object.assign({}, this.metadata, {json: true});
+    const metadata = Object.assign({}, this.metadata, { json: true });
 
     for (let i = 0; i < metadata.row_groups.length; i++) {
       const rowGroup = metadata.row_groups[i];
       for (let j = 0; j < rowGroup.columns.length; j++) {
         const column = rowGroup.columns[j];
-          if (column.offsetIndex instanceof Promise) {
-            column.offsetIndex = await column.offsetIndex;
-          }
-          if (column.columnIndex instanceof Promise) {
-            column.columnIndex = await column.columnIndex;
-          }
+        if (column.offsetIndex instanceof Promise) {
+          column.offsetIndex = await column.offsetIndex;
+        }
+        if (column.columnIndex instanceof Promise) {
+          column.columnIndex = await column.columnIndex;
+        }
       }
     }
 
-    return JSON.stringify(metadata,replacer,indent);
+    return JSON.stringify(metadata, replacer, indent);
   }
 
   /**
@@ -380,7 +393,6 @@ export class ParquetReader {
   decodePages(buffer: Buffer, opts: Options) {
     return decodePages(buffer, opts);
   }
-
 }
 
 /**
@@ -395,16 +407,16 @@ export class ParquetEnvelopeReader {
   readFn: (offset: number, length: number, file?: string) => Promise<Buffer>;
   close: () => unknown;
   id: number;
-  fileSize: Function | number;
+  fileSize: number | (() => Promise<number>);
   default_dictionary_size: number;
   metadata?: FileMetaDataExt;
-  schema?: parquet_schema.ParquetSchema
+  schema?: parquet_schema.ParquetSchema;
 
   static async openFile(filePath: string | Buffer | URL, options?: BufferReaderOptions) {
-    let fileStat = await parquet_util.fstat(filePath);
-    let fileDescriptor = await parquet_util.fopen(filePath);
+    const fileStat = await parquet_util.fstat(filePath);
+    const fileDescriptor = await parquet_util.fopen(filePath);
 
-    let readFn = (offset: number, length: number, file?: string) => {
+    const readFn = (offset: number, length: number, file?: string) => {
       if (file) {
         return Promise.reject('external references are not supported');
       }
@@ -412,66 +424,120 @@ export class ParquetEnvelopeReader {
       return parquet_util.fread(fileDescriptor, offset, length);
     };
 
-    let closeFn = parquet_util.fclose.bind(undefined, fileDescriptor);
+    const closeFn = parquet_util.fclose.bind(undefined, fileDescriptor);
 
     return new ParquetEnvelopeReader(readFn, closeFn, fileStat.size, options);
   }
 
   static async openBuffer(buffer: Buffer, options?: BufferReaderOptions) {
-    let readFn = (offset: number, length: number, file?: string) => {
+    const readFn = (offset: number, length: number, file?: string) => {
       if (file) {
         return Promise.reject('external references are not supported');
       }
 
-      return Promise.resolve(buffer.slice(offset, offset + length));
+      return Promise.resolve(buffer.subarray(offset, offset + length));
     };
 
-    let closeFn = () => ({});
+    const closeFn = () => ({});
     return new ParquetEnvelopeReader(readFn, closeFn, buffer.length, options);
   }
 
   static async openS3(client: ClientS3, params: ClientParameters, options?: BufferReaderOptions) {
-    let fileStat = async () => client.headObject(params).promise().then((d: { ContentLength: number }) => d.ContentLength);
+    const fileStat = async () =>
+      client
+        .headObject(params)
+        .promise()
+        .then((d: { ContentLength: number }) => d.ContentLength);
 
-    let readFn = async (offset: number, length: number, file?: string) => {
+    const readFn = async (offset: number, length: number, file?: string) => {
       if (file) {
         return Promise.reject('external references are not supported');
       }
 
-      let Range = `bytes=${offset}-${offset+length-1}`;
-      let res = await client.getObject(Object.assign({Range}, params)).promise();
+      const Range = `bytes=${offset}-${offset + length - 1}`;
+      const res = await client.getObject(Object.assign({ Range }, params)).promise();
       return Promise.resolve(res.Body);
     };
 
-    let closeFn = () => ({});
+    const closeFn = () => ({});
 
     return new ParquetEnvelopeReader(readFn, closeFn, fileStat, options);
+  }
+
+  static async openS3v3(client: S3Client, params: any, options: any) {
+    const fileStat = async () => {
+      try {
+        const headObjectCommand = await client.send(new HeadObjectCommand(params));
+        if (headObjectCommand.ContentLength === undefined) {
+          throw new Error('Content Length is undefined!');
+        }
+        return Promise.resolve(headObjectCommand.ContentLength);
+      } catch (e: any) {
+        // having params match command names makes e.message clear to user
+        return Promise.reject('rejected headObjectCommand: ' + e.message);
+      }
+    };
+
+    const readFn = async (offset: number, length: number, file: string | undefined): Promise<Buffer> => {
+      if (file) {
+        return Promise.reject('external references are not supported');
+      }
+      const Range = `bytes=${offset}-${offset + length - 1}`;
+      const input = { ...{ Range }, ...params };
+      const response = await client.send(new GetObjectCommand(input));
+
+      const body = response.Body;
+      if (body) {
+        return ParquetEnvelopeReader.streamToBuffer(body);
+      }
+      return Buffer.of();
+    };
+
+    const closeFn = () => ({});
+
+    return new ParquetEnvelopeReader(readFn, closeFn, fileStat, options);
+  }
+
+  static async streamToBuffer(body: any): Promise<Buffer> {
+    const blob = body as Blob;
+    if (blob.arrayBuffer !== undefined) {
+      const arrayBuffer = await blob.arrayBuffer();
+      const uint8Array: Uint8Array = new Uint8Array(arrayBuffer);
+      return Buffer.from(uint8Array);
+    }
+
+    //Assumed to be a Readable like object
+    const readable = body as Readable;
+    return await new Promise((resolve, reject) => {
+      const chunks: Uint8Array[] = [];
+      readable.on('data', (chunk) => chunks.push(chunk));
+      readable.on('error', reject);
+      readable.on('end', () => resolve(Buffer.concat(chunks)));
+    });
   }
 
   static async openUrl(url: Parameter | URL | string, options?: BufferReaderOptions) {
     let params: Parameter;
     if (typeof url === 'string') params = { url };
-    else if(url instanceof URL) params = { url: url.toString() }
+    else if (url instanceof URL) params = { url: url.toString() };
     else params = url;
 
-    if (!params.url)
-      throw new Error('URL missing');
+    if (!params.url) throw new Error('URL missing');
 
     const baseArr = params.url.split('/');
-    const base = baseArr.slice(0, baseArr.length-1).join('/')+'/';
+    const base = baseArr.slice(0, baseArr.length - 1).join('/') + '/';
 
-    let defaultHeaders = params.headers || {};
+    const defaultHeaders = params.headers || {};
 
-    let filesize = async () => {
-
+    const filesize = async (): Promise<number> => {
       const { headers } = await fetch(params.url);
-      return headers.get('Content-Length');
+      return Number(headers.get('Content-Length')) || 0;
     };
 
-    let readFn = async (offset: number, length: number, file?: string) => {
-      let url = file ? base+file : params.url;
-      let range = `bytes=${offset}-${offset+length-1}`;
-      let headers = Object.assign({}, defaultHeaders, {range});
+    const readFn = async (offset: number, length: number, file?: string) => {
+      const url = file ? base + file : params.url;
+      const range = `bytes=${offset}-${offset + length - 1}`;
+      const headers = Object.assign({}, defaultHeaders, { range });
       const response = await fetch(url, { headers });
       const arrayBuffer = await response.arrayBuffer();
       const buffer = Buffer.from(arrayBuffer);
@@ -479,12 +545,18 @@ export class ParquetEnvelopeReader {
       return buffer;
     };
 
-    let closeFn = () => ({});
+    const closeFn = () => ({});
 
     return new ParquetEnvelopeReader(readFn, closeFn, filesize, options);
   }
 
-  constructor(readFn: (offset: number, length: number, file?: string) => Promise<Buffer> , closeFn: () => unknown, fileSize: Function | number, options?: BufferReaderOptions, metadata?: FileMetaDataExt) {
+  constructor(
+    readFn: (offset: number, length: number, file?: string) => Promise<Buffer>,
+    closeFn: () => unknown,
+    fileSize: number | (() => Promise<number>),
+    options?: BufferReaderOptions,
+    metadata?: FileMetaDataExt
+  ) {
     options = options || {};
     this.readFn = readFn;
     this.id = ++ParquetEnvelopeReaderIdCounter;
@@ -504,15 +576,17 @@ export class ParquetEnvelopeReader {
 
   readHeader() {
     return this.read(0, PARQUET_MAGIC.length).then((buf: Buffer) => {
-
       if (buf.toString() != PARQUET_MAGIC) {
-        throw 'not valid parquet file'
+        throw 'not valid parquet file';
       }
     });
   }
 
   // Helper function to get the column object for a particular path and row_group
-  getColumn(path: string | parquet_thrift.ColumnChunk, row_group: RowGroupExt | number | string | null) : ColumnChunkExt {
+  getColumn(
+    path: string | parquet_thrift.ColumnChunk,
+    row_group: RowGroupExt | number | string | null
+  ): ColumnChunkExt {
     let column;
     let parsedRowGroup: parquet_thrift.RowGroup | undefined;
     if (!isNaN(Number(row_group))) {
@@ -523,9 +597,9 @@ export class ParquetEnvelopeReader {
 
     if (typeof path === 'string') {
       if (!parsedRowGroup) {
-       throw `Missing RowGroup ${row_group}`;
+        throw `Missing RowGroup ${row_group}`;
       }
-        column = parsedRowGroup.columns.find(d => d.meta_data!.path_in_schema.join(',') === path);
+      column = parsedRowGroup.columns.find((d) => d.meta_data!.path_in_schema.join(',') === path);
 
       if (!column) {
         throw `Column ${path} Not Found`;
@@ -536,21 +610,25 @@ export class ParquetEnvelopeReader {
     return column;
   }
 
-  getAllColumnChunkDataFor(paths: Array<string>, row_groups?: Array<RowGroupExt>) {
+  getAllColumnChunkDataFor(paths: string[], row_groups?: RowGroupExt[]) {
     if (!row_groups) {
       row_groups = this.metadata!.row_groups;
     }
 
     return row_groups.flatMap((rowGroup, index) =>
-              paths.map(columnName => ({
-                  rowGroupIndex: index,
-                  column: this.getColumn(columnName, rowGroup)
-              }))
-            )
+      paths.map((columnName) => ({
+        rowGroupIndex: index,
+        column: this.getColumn(columnName, rowGroup),
+      }))
+    );
   }
 
-  readOffsetIndex(path: string | ColumnChunkExt, row_group: RowGroupExt | number | null, opts: Options): Promise<parquet_thrift.OffsetIndex> {
-    let column = this.getColumn(path, row_group);
+  readOffsetIndex(
+    path: string | ColumnChunkExt,
+    row_group: RowGroupExt | number | null,
+    opts: Options
+  ): Promise<parquet_thrift.OffsetIndex> {
+    const column = this.getColumn(path, row_group);
     if (column.offsetIndex) {
       return Promise.resolve(column.offsetIndex);
     } else if (!column.offset_index_offset || !column.offset_index_length) {
@@ -558,9 +636,9 @@ export class ParquetEnvelopeReader {
     }
 
     const data = this.read(+column.offset_index_offset, column.offset_index_length).then((data: Buffer) => {
-      let offset_index = new parquet_thrift.OffsetIndex();
+      const offset_index = new parquet_thrift.OffsetIndex();
       parquet_util.decodeThrift(offset_index, data);
-      Object.defineProperty(offset_index,'column', {value: column, enumerable: false});
+      Object.defineProperty(offset_index, 'column', { value: column, enumerable: false });
       return offset_index;
     });
     if (opts?.cache) {
@@ -569,26 +647,30 @@ export class ParquetEnvelopeReader {
     return data;
   }
 
-  readColumnIndex(path: string | ColumnChunkExt, row_group: RowGroupExt | number, opts: Options): Promise<parquet_thrift.ColumnIndex> {
-    let column = this.getColumn(path, row_group);
+  readColumnIndex(
+    path: string | ColumnChunkExt,
+    row_group: RowGroupExt | number,
+    opts: Options
+  ): Promise<parquet_thrift.ColumnIndex> {
+    const column = this.getColumn(path, row_group);
     if (column.columnIndex) {
       return Promise.resolve(column.columnIndex);
     } else if (!column.column_index_offset) {
       return Promise.reject(new Error('Column Index Missing'));
     }
 
-    const data = this.read(+column.column_index_offset, (column.column_index_length as number)).then((buf: Buffer) => {
-      let column_index = new parquet_thrift.ColumnIndex();
+    const data = this.read(+column.column_index_offset, column.column_index_length as number).then((buf: Buffer) => {
+      const column_index = new parquet_thrift.ColumnIndex();
       parquet_util.decodeThrift(column_index, buf);
       Object.defineProperty(column_index, 'column', { value: column });
 
       // decode the statistics values
       const field = this.schema!.findField(column.meta_data!.path_in_schema);
       if (column_index.max_values) {
-        column_index.max_values = column_index.max_values.map(max_value => decodeStatisticsValue(max_value, field));
+        column_index.max_values = column_index.max_values.map((max_value) => decodeStatisticsValue(max_value, field));
       }
       if (column_index.min_values) {
-        column_index.min_values = column_index.min_values.map(min_value => decodeStatisticsValue(min_value, field));
+        column_index.min_values = column_index.min_values.map((min_value) => decodeStatisticsValue(min_value, field));
       }
       return column_index;
     });
@@ -599,39 +681,46 @@ export class ParquetEnvelopeReader {
     return data;
   }
 
-  async readPage(column: ColumnChunkExt, page: parquet_thrift.PageLocation | number, records: Array<Record<string, unknown>>, opts: Options) {
-    column = Object.assign({},column);
-    column.meta_data = Object.assign({},column.meta_data);
+  async readPage(
+    column: ColumnChunkExt,
+    page: parquet_thrift.PageLocation | number,
+    records: Record<string, unknown>[],
+    opts: Options
+  ) {
+    column = Object.assign({}, column);
+    column.meta_data = Object.assign({}, column.meta_data);
 
     if (page instanceof parquet_thrift.PageLocation && page.offset !== undefined) {
       if (isNaN(Number(page.offset)) || isNaN(page.compressed_page_size)) {
         throw Error('page offset and/or size missing');
       }
       column.meta_data.data_page_offset = parquet_util.cloneInteger(page.offset);
-      column.meta_data.total_compressed_size =  new Int64(page.compressed_page_size);
+      column.meta_data.total_compressed_size = new Int64(page.compressed_page_size);
     } else {
       const offsetIndex = await this.readOffsetIndex(column, null, opts);
       column.meta_data.data_page_offset = parquet_util.cloneInteger(offsetIndex.page_locations[page as number].offset);
-      column.meta_data.total_compressed_size =  new Int64(offsetIndex.page_locations[page as number].compressed_page_size);
+      column.meta_data.total_compressed_size = new Int64(
+        offsetIndex.page_locations[page as number].compressed_page_size
+      );
     }
     const chunk = await this.readColumnChunk(this.schema!, column);
-    Object.defineProperty(chunk,'column', {value: column});
-    let data = {
-      columnData: {[chunk.column!.meta_data!.path_in_schema.join(',')]: chunk}
+    Object.defineProperty(chunk, 'column', { value: column });
+    const data = {
+      columnData: { [chunk.column!.meta_data!.path_in_schema.join(',')]: chunk },
     };
 
     return parquet_shredder.materializeRecords(this.schema!, data, records);
   }
 
-  async readRowGroup(schema: parquet_schema.ParquetSchema, rowGroup: RowGroupExt, columnList: Array<Array<unknown>>) {
-    var buffer: parquet_shredder.RecordBuffer = {
+  async readRowGroup(schema: parquet_schema.ParquetSchema, rowGroup: RowGroupExt, columnList: unknown[][]) {
+    const buffer: parquet_shredder.RecordBuffer = {
       rowCount: +rowGroup.num_rows,
       columnData: {},
       pageRowCount: 0,
-      pages: {}
+      pages: {},
     };
 
-    for (let colChunk of rowGroup.columns) {
+    for (const colChunk of rowGroup.columns) {
       const colMetadata = colChunk.meta_data;
       const colKey = colMetadata!.path_in_schema;
 
@@ -646,45 +735,43 @@ export class ParquetEnvelopeReader {
   }
 
   async readColumnChunk(schema: parquet_schema.ParquetSchema, colChunk: ColumnChunkExt, opts?: Options) {
-    let metadata = colChunk.meta_data!
-    let field = schema.findField(metadata.path_in_schema);
-    let type = parquet_util.getThriftEnum(
-        parquet_thrift.Type,
-        metadata.type);
+    const metadata = colChunk.meta_data!;
+    const field = schema.findField(metadata.path_in_schema);
+    const type = parquet_util.getThriftEnum(parquet_thrift.Type, metadata.type);
 
-    let compression = parquet_util.getThriftEnum(
-        parquet_thrift.CompressionCodec,
-        metadata.codec);
+    const compression = parquet_util.getThriftEnum(parquet_thrift.CompressionCodec, metadata.codec);
 
-    let pagesOffset = +metadata.data_page_offset;
+    const pagesOffset = +metadata.data_page_offset;
     let pagesSize = +metadata.total_compressed_size;
 
     if (!colChunk.file_path) {
       pagesSize = Math.min((this.fileSize as number) - pagesOffset, +metadata.total_compressed_size);
     }
 
-    opts = Object.assign({},opts, {
+    opts = Object.assign({}, opts, {
       type: type,
       rLevelMax: field.rLevelMax,
       dLevelMax: field.dLevelMax,
       compression: compression,
       column: field,
-      num_values: metadata.num_values
+      num_values: metadata.num_values,
     });
 
-    if (metadata.dictionary_page_offset) {
-      const offset = +metadata.dictionary_page_offset;
+    // If this exists and is greater than zero then we need to have an offset
+    if (metadata.dictionary_page_offset && +metadata.dictionary_page_offset > 0) {
+      const offset: number = +metadata.dictionary_page_offset;
       const size = Math.min(+this.fileSize - offset, this.default_dictionary_size);
 
       await this.read(offset, size, colChunk.file_path).then(async (buffer: Buffer) => {
-        await decodePage({offset: 0, buffer, size: buffer.length}, opts!).then(dict => {
-          opts!.dictionary = opts!.dictionary || dict.dictionary as number[];
-        })
-      })
-
+        await decodePage({ offset: 0, buffer, size: buffer.length }, opts!).then((dict) => {
+          opts!.dictionary = opts!.dictionary || (dict.dictionary as number[]);
+        });
+      });
     }
 
-    return this.read(pagesOffset, pagesSize, colChunk.file_path).then((pagesBuf: Buffer) => decodePages(pagesBuf, opts!));
+    return this.read(pagesOffset, pagesSize, colChunk.file_path).then((pagesBuf: Buffer) =>
+      decodePages(pagesBuf, opts!)
+    );
   }
 
   async readFooter() {
@@ -692,46 +779,57 @@ export class ParquetEnvelopeReader {
       this.fileSize = await this.fileSize();
     }
 
-    let trailerLen = PARQUET_MAGIC.length + 4;
+    const trailerLen = PARQUET_MAGIC.length + 4;
 
-    let trailerBuf = await this.read((this.fileSize as number) - trailerLen, trailerLen);
+    const offset = (this.fileSize as number) - trailerLen;
+    const trailerBuf = await this.read(offset, trailerLen);
 
-    if (trailerBuf.slice(4).toString() != PARQUET_MAGIC) {
+    if (trailerBuf.subarray(4).toString() != PARQUET_MAGIC) {
       throw 'not a valid parquet file';
     }
 
-    let metadataSize = trailerBuf.readUInt32LE(0);
-    let metadataOffset = (this.fileSize as number) - metadataSize - trailerLen;
+    const metadataSize = trailerBuf.readUInt32LE(0);
+    const metadataOffset = (this.fileSize as number) - metadataSize - trailerLen;
     if (metadataOffset < PARQUET_MAGIC.length) {
       throw 'invalid metadata size';
     }
 
-    let metadataBuf = await this.read(metadataOffset, metadataSize);
-    let metadata = new parquet_thrift.FileMetaData();
+    const metadataBuf = await this.read(metadataOffset, metadataSize);
+    const metadata = new parquet_thrift.FileMetaData();
     parquet_util.decodeThrift(metadata, metadataBuf);
     return metadata;
   }
-
 }
 
 /**
  * Decode a consecutive array of data using one of the parquet encodings
  */
-function decodeValues(type: string, encoding: ParquetCodec, cursor: Cursor, count: number, opts: Options | {bitWidth: number}) {
+function decodeValues(
+  type: string,
+  encoding: ParquetCodec,
+  cursor: Cursor,
+  count: number,
+  opts: Options | { bitWidth: number }
+) {
   if (!(encoding in parquet_codec)) {
-    throw 'invalid encoding: ' + encoding;
+      throw 'invalid encoding: ' + encoding + JSON.stringify(parquet_codec);
   }
 
   return parquet_codec[encoding].decodeValues(type, cursor, count, opts as Options);
 }
-
 
 function decodeStatisticsValue(value: any, column: ParquetField | Options) {
   if (value === null || !value.length) {
     return undefined;
   }
   if (!column.primitiveType!.includes('BYTE_ARRAY')) {
-    value = decodeValues(column.primitiveType!,'PLAIN',{buffer: Buffer.from(value), offset: 0}, 1, column as Options);
+    value = decodeValues(
+      column.primitiveType!,
+      'PLAIN',
+      { buffer: Buffer.from(value), offset: 0 },
+      1,
+      column as Options
+    );
     if (value.length === 1) value = value[0];
   }
 
@@ -764,30 +862,33 @@ async function decodePage(cursor: Cursor, opts: Options): Promise<PageData> {
   const pageHeader = new NewPageHeader();
 
   const headerOffset = cursor.offset;
-  const headerSize = parquet_util.decodeThrift(pageHeader, cursor.buffer.slice(cursor.offset));
+  const headerSize = parquet_util.decodeThrift(pageHeader, cursor.buffer.subarray(cursor.offset));
   cursor.offset += headerSize;
 
-  const pageType = parquet_util.getThriftEnum(
-      parquet_thrift.PageType,
-      pageHeader.type);
+  const pageType = parquet_util.getThriftEnum(parquet_thrift.PageType, pageHeader.type);
 
   switch (pageType) {
     case 'DATA_PAGE':
       if (!opts.rawStatistics) {
-        pageHeader.data_page_header!.statistics = decodeStatistics(pageHeader.data_page_header!.statistics!, opts.column!);
+        pageHeader.data_page_header!.statistics = decodeStatistics(
+          pageHeader.data_page_header!.statistics!,
+          opts.column!
+        );
       }
       page = await decodeDataPage(cursor, pageHeader, opts);
       break;
     case 'DATA_PAGE_V2':
       if (!opts.rawStatistics) {
-        pageHeader.data_page_header_v2!.statistics = decodeStatistics(pageHeader.data_page_header_v2!.statistics!, opts.column!);
+        pageHeader.data_page_header_v2!.statistics = decodeStatistics(
+          pageHeader.data_page_header_v2!.statistics!,
+          opts.column!
+        );
       }
       page = await decodeDataPageV2(cursor, pageHeader, opts);
       break;
     case 'DICTIONARY_PAGE':
-      const dict = await decodeDictionaryPage(cursor, pageHeader, opts)
       page = {
-        dictionary: dict
+        dictionary: await decodeDictionaryPage(cursor, pageHeader, opts),
       };
       break;
     default:
@@ -801,21 +902,20 @@ async function decodePage(cursor: Cursor, opts: Options): Promise<PageData> {
   return page;
 }
 
-
 async function decodePages(buffer: Buffer, opts: Options) {
   opts = opts || {};
-  let cursor = {
+  const cursor = {
     buffer: buffer,
     offset: 0,
-    size: buffer.length
+    size: buffer.length,
   };
 
-  let data: PageData = {
+  const data: PageData = {
     rlevels: [],
     dlevels: [],
     values: [],
     pageHeaders: [],
-    count: 0
+    count: 0,
   };
 
   while (cursor.offset < cursor.size && (!opts.num_values || data.dlevels!.length < opts.num_values)) {
@@ -831,15 +931,15 @@ async function decodePages(buffer: Buffer, opts: Options) {
     // not (PLAIN for example).
 
     if (opts.dictionary && pageData.useDictionary) {
-      pageData.values = pageData.values!.map(d => opts.dictionary![d]);
+      pageData.values = pageData.values!.map((d) => opts.dictionary![d]);
     }
 
-    let length = pageData.rlevels != undefined ? pageData.rlevels.length : 0;
+    const length = pageData.rlevels != undefined ? pageData.rlevels.length : 0;
 
     for (let i = 0; i < length; i++) {
       data.rlevels!.push(pageData.rlevels![i]);
       data.dlevels!.push(pageData.dlevels![i]);
-      let value = pageData.values![i];
+      const value = pageData.values![i];
       if (value !== undefined) {
         data.values!.push(value);
       }
@@ -856,21 +956,22 @@ async function decodeDictionaryPage(cursor: Cursor, header: parquet_thrift.PageH
 
   let dictCursor = {
     offset: 0,
-    buffer: cursor.buffer.slice(cursor.offset,cursorEnd),
-    size: cursorEnd - cursor.offset
+    buffer: cursor.buffer.subarray(cursor.offset, cursorEnd),
+    size: cursorEnd - cursor.offset,
   };
 
   cursor.offset = cursorEnd;
 
   if (opts.compression && opts.compression !== 'UNCOMPRESSED') {
-    let valuesBuf = await parquet_compression.inflate(
-        opts.compression,
-        dictCursor.buffer.slice(dictCursor.offset,cursorEnd));
+    const valuesBuf = await parquet_compression.inflate(
+      opts.compression,
+      dictCursor.buffer.subarray(dictCursor.offset, cursorEnd)
+    );
 
     dictCursor = {
       buffer: valuesBuf,
       offset: 0,
-      size: valuesBuf.length
+      size: valuesBuf.length,
     };
   }
 
@@ -878,7 +979,7 @@ async function decodeDictionaryPage(cursor: Cursor, header: parquet_thrift.PageH
     opts.column!.primitiveType!,
     opts.column!.encoding!,
     dictCursor,
-    (header.dictionary_page_header!).num_values,
+    header.dictionary_page_header!.num_values,
     { ...opts, ...opts.column }
   )
 }
@@ -888,82 +989,64 @@ async function decodeDataPage(cursor: Cursor, header: parquet_thrift.PageHeader,
 
   const dataPageHeader = header.data_page_header!;
 
-  let valueCount = dataPageHeader.num_values;
-  let valueEncoding = parquet_util.getThriftEnum(
-      parquet_thrift.Encoding,
-      dataPageHeader.encoding);
+  const valueCount = dataPageHeader.num_values;
+  const valueEncoding = parquet_util.getThriftEnum(parquet_thrift.Encoding, dataPageHeader.encoding);
 
   let valuesBufCursor = cursor;
   if (opts.compression && opts.compression !== 'UNCOMPRESSED') {
-    let valuesBuf = await parquet_compression.inflate(
-        opts.compression,
-        cursor.buffer.slice(cursor.offset, cursorEnd));
+    const valuesBuf = await parquet_compression.inflate(
+      opts.compression,
+      cursor.buffer.subarray(cursor.offset, cursorEnd)
+    );
 
     valuesBufCursor = {
       buffer: valuesBuf,
       offset: 0,
-      size: valuesBuf.length
+      size: valuesBuf.length,
     };
   }
 
-
-
   /* read repetition levels */
-  let rLevelEncoding = parquet_util.getThriftEnum(
-      parquet_thrift.Encoding,
-      dataPageHeader.repetition_level_encoding);
+  const rLevelEncoding = parquet_util.getThriftEnum(parquet_thrift.Encoding, dataPageHeader.repetition_level_encoding);
 
   let rLevels = new Array(valueCount);
   if (opts.rLevelMax! > 0) {
-    rLevels = decodeValues(
-        PARQUET_RDLVL_TYPE,
-        rLevelEncoding as ParquetCodec,
-        valuesBufCursor,
-        valueCount,
-        { bitWidth: parquet_util.getBitWidth(opts.rLevelMax!) });
+    rLevels = decodeValues(PARQUET_RDLVL_TYPE, rLevelEncoding as ParquetCodec, valuesBufCursor, valueCount, {
+      bitWidth: parquet_util.getBitWidth(opts.rLevelMax!),
+    });
   } else {
     rLevels.fill(0);
   }
 
   /* read definition levels */
-  let dLevelEncoding = parquet_util.getThriftEnum(
-      parquet_thrift.Encoding,
-      dataPageHeader.definition_level_encoding);
+  const dLevelEncoding = parquet_util.getThriftEnum(parquet_thrift.Encoding, dataPageHeader.definition_level_encoding);
 
   let dLevels = new Array(valueCount);
   if (opts.dLevelMax! > 0) {
-    dLevels = decodeValues(
-        PARQUET_RDLVL_TYPE,
-        dLevelEncoding as ParquetCodec,
-        valuesBufCursor,
-        valueCount,
-        { bitWidth: parquet_util.getBitWidth(opts.dLevelMax!) });
+    dLevels = decodeValues(PARQUET_RDLVL_TYPE, dLevelEncoding as ParquetCodec, valuesBufCursor, valueCount, {
+      bitWidth: parquet_util.getBitWidth(opts.dLevelMax!),
+    });
   } else {
     dLevels.fill(0);
   }
 
   /* read values */
   let valueCountNonNull = 0;
-  for (let dlvl of dLevels) {
+  for (const dlvl of dLevels) {
     if (dlvl === opts.dLevelMax) {
       ++valueCountNonNull;
     }
   }
 
-  let values = decodeValues(
-      opts.type!,
-      valueEncoding as ParquetCodec,
-      valuesBufCursor,
-      valueCountNonNull,
-      {
-        typeLength: opts.column!.typeLength!,
-        bitWidth: opts.column!.typeLength!,
-        disableEnvelope: opts.column!.disableEnvelope,
-        originalType: opts.column!.originalType,
-        precision: opts.column!.precision,
-        scale: opts.column!.scale,
-        name: opts.column!.name
-      });
+  const values = decodeValues(opts.type!, valueEncoding as ParquetCodec, valuesBufCursor, valueCountNonNull, {
+    typeLength: opts.column!.typeLength!,
+    bitWidth: opts.column!.typeLength!,
+    disableEnvelope: opts.column!.disableEnvelope,
+    originalType: opts.column!.originalType,
+    precision: opts.column!.precision,
+    scale: opts.column!.scale,
+    name: opts.column!.name,
+  });
 
   cursor.offset = cursorEnd;
 
@@ -972,7 +1055,7 @@ async function decodeDataPage(cursor: Cursor, header: parquet_thrift.PageHeader,
     rlevels: rLevels,
     values: values,
     count: valueCount,
-    useDictionary: valueEncoding === 'PLAIN_DICTIONARY' || valueEncoding === 'RLE_DICTIONARY'
+    useDictionary: valueEncoding === 'PLAIN_DICTIONARY' || valueEncoding === 'RLE_DICTIONARY',
   };
 }
 
@@ -982,22 +1065,15 @@ async function decodeDataPageV2(cursor: Cursor, header: parquet_thrift.PageHeade
 
   const valueCount = dataPageHeaderV2.num_values;
   const valueCountNonNull = valueCount - dataPageHeaderV2.num_nulls;
-  const valueEncoding = parquet_util.getThriftEnum(
-      parquet_thrift.Encoding,
-      dataPageHeaderV2.encoding);
+  const valueEncoding = parquet_util.getThriftEnum(parquet_thrift.Encoding, dataPageHeaderV2.encoding);
 
   /* read repetition levels */
   let rLevels = new Array(valueCount);
   if (opts.rLevelMax! > 0) {
-    rLevels = decodeValues(
-        PARQUET_RDLVL_TYPE,
-        PARQUET_RDLVL_ENCODING,
-        cursor,
-        valueCount,
-        {
-          bitWidth: parquet_util.getBitWidth(opts.rLevelMax!),
-          disableEnvelope: true
-        });
+    rLevels = decodeValues(PARQUET_RDLVL_TYPE, PARQUET_RDLVL_ENCODING, cursor, valueCount, {
+      bitWidth: parquet_util.getBitWidth(opts.rLevelMax!),
+      disableEnvelope: true,
+    });
   } else {
     rLevels.fill(0);
   }
@@ -1005,15 +1081,10 @@ async function decodeDataPageV2(cursor: Cursor, header: parquet_thrift.PageHeade
   /* read definition levels */
   let dLevels = new Array(valueCount);
   if (opts.dLevelMax! > 0) {
-    dLevels = decodeValues(
-        PARQUET_RDLVL_TYPE,
-        PARQUET_RDLVL_ENCODING,
-        cursor,
-        valueCount,
-        {
-          bitWidth: parquet_util.getBitWidth(opts.dLevelMax!),
-          disableEnvelope: true
-        });
+    dLevels = decodeValues(PARQUET_RDLVL_TYPE, PARQUET_RDLVL_ENCODING, cursor, valueCount, {
+      bitWidth: parquet_util.getBitWidth(opts.dLevelMax!),
+      disableEnvelope: true,
+    });
   } else {
     dLevels.fill(0);
   }
@@ -1022,45 +1093,41 @@ async function decodeDataPageV2(cursor: Cursor, header: parquet_thrift.PageHeade
   let valuesBufCursor = cursor;
 
   if (dataPageHeaderV2.is_compressed) {
-    let valuesBuf = await parquet_compression.inflate(
-        opts.compression!,
-        cursor.buffer.slice(cursor.offset, cursorEnd));
+    const valuesBuf = await parquet_compression.inflate(
+      opts.compression!,
+      cursor.buffer.subarray(cursor.offset, cursorEnd)
+    );
 
     valuesBufCursor = {
       buffer: valuesBuf,
       offset: 0,
-      size: valuesBuf.length
+      size: valuesBuf.length,
     };
 
     cursor.offset = cursorEnd;
   }
 
-  let values = decodeValues(
-      opts.type!,
-      valueEncoding as ParquetCodec,
-      valuesBufCursor,
-      valueCountNonNull,
-      {
-        bitWidth: opts.column!.typeLength!,
-        ...opts.column!
-      });
+  const values = decodeValues(opts.type!, valueEncoding as ParquetCodec, valuesBufCursor, valueCountNonNull, {
+    bitWidth: opts.column!.typeLength!,
+    ...opts.column!,
+  });
 
   return {
     dlevels: dLevels,
     rlevels: rLevels,
     values: values,
     count: valueCount,
-    useDictionary: valueEncoding === 'PLAIN_DICTIONARY' || valueEncoding === 'RLE_DICTIONARY'
+    useDictionary: valueEncoding === 'PLAIN_DICTIONARY' || valueEncoding === 'RLE_DICTIONARY',
   };
 }
 
-function decodeSchema(schemaElements: Array<parquet_thrift.SchemaElement>) {
-   let schema: SchemaDefinition | FieldDefinition = {};
-  schemaElements.forEach(schemaElement => {
-
-    let repetitionType = parquet_util.getThriftEnum(
-        parquet_thrift.FieldRepetitionType,
-        schemaElement.repetition_type);
+function decodeSchema(schemaElements: parquet_thrift.SchemaElement[]) {
+  let schema: SchemaDefinition | FieldDefinition = {};
+  schemaElements.forEach((schemaElement) => {
+    const repetitionType = parquet_util.getThriftEnum(
+      parquet_thrift.FieldRepetitionType,
+      schemaElement.repetition_type
+    );
 
     let optional = false;
     let repeated = false;
@@ -1073,35 +1140,34 @@ function decodeSchema(schemaElements: Array<parquet_thrift.SchemaElement>) {
       case 'REPEATED':
         repeated = true;
         break;
-    };
+    }
 
     if (schemaElement.num_children != undefined && schemaElement.num_children > 0) {
       (schema as SchemaDefinition)[schemaElement.name] = {
         optional: optional,
         repeated: repeated,
-        fields: Object.create({},{
-          /* define parent and num_children as non-enumerable */
-          parent: {
-            value: schema,
-            enumerable: false
-          },
-          num_children: {
-            value: schemaElement.num_children,
-            enumerable: false
+        fields: Object.create(
+          {},
+          {
+            /* define parent and num_children as non-enumerable */
+            parent: {
+              value: schema,
+              enumerable: false,
+            },
+            num_children: {
+              value: schemaElement.num_children,
+              enumerable: false,
+            },
           }
-        })
+        ),
       };
       /* move the schema pointer to the children */
       schema = (schema as SchemaDefinition)[schemaElement.name].fields as SchemaDefinition;
     } else {
-      let logicalType = parquet_util.getThriftEnum(
-          parquet_thrift.Type,
-          schemaElement.type);
+      let logicalType = parquet_util.getThriftEnum(parquet_thrift.Type, schemaElement.type);
 
       if (schemaElement.converted_type != null) {
-        logicalType = parquet_util.getThriftEnum(
-            parquet_thrift.ConvertedType,
-            schemaElement.converted_type);
+        logicalType = parquet_util.getThriftEnum(parquet_thrift.ConvertedType, schemaElement.converted_type);
       }
 
       (schema as SchemaDefinition)[schemaElement.name] = {
@@ -1110,7 +1176,7 @@ function decodeSchema(schemaElements: Array<parquet_thrift.SchemaElement>) {
         optional: optional,
         repeated: repeated,
         scale: schemaElement.scale,
-        precision: schemaElement.precision
+        precision: schemaElement.precision,
       };
     }
 
