@@ -4,9 +4,11 @@ const assert = chai.assert;
 const parquet = require('../parquet');
 const path = require('path');
 
-describe('dictionary encoding', async function() {
-  it('should read uncompressed dictionary from spark', async function() {
-    let reader =  await parquet.ParquetReader.openFile(path.resolve(__dirname,'test-files/spark-uncompressed-dict.parquet'));
+describe('dictionary encoding', function () {
+  it('should read uncompressed dictionary from spark', async function () {
+    let reader = await parquet.ParquetReader.openFile(
+      path.resolve(__dirname, 'test-files/spark-uncompressed-dict.parquet')
+    );
     let cursor = reader.getCursor();
     let records = [];
 
@@ -14,6 +16,9 @@ describe('dictionary encoding', async function() {
       records.push(await cursor.next());
     }
 
-    assert.deepEqual(records.map(d => d.name),['apples','oranges','kiwi','banana','apples']);
+    assert.deepEqual(
+      records.map((d) => d.name),
+      ['apples', 'oranges', 'kiwi', 'banana', 'apples']
+    );
   });
 });
