@@ -1,4 +1,4 @@
-const { setupServer } = require("msw/node");
-const handlers = require("./handlers");
+const { setupServer } = require('msw/node');
+const handlers = require('./handlers');
 
 module.exports = setupServer(...handlers);
